requirements.txt
#torch==2.1.1 
#torchvision==0.16.1
pandas==2.1.3
numpy==1.26.2
Pillow==10.1.0
matplotlib==3.8.2
scikit-learn==1.3.2
kaggle==1.5.16